// src/clients/http_client/types.ts
import { Method } from "@shop0/network";
import { Headers } from "node-fetch";

export type HeaderParams = Record<string, string | number>;

export enum DataType {
  JSON = "application/json", // eslint-disable-line @shop0/typescript/prefer-pascal-case-enums
  GraphQL = "application/graphql", // eslint-disable-line @shop0/typescript/prefer-pascal-case-enums
  URLEncoded = "application/x-www-form-urlencoded", // eslint-disable-line @shop0/typescript/prefer-pascal-case-enums
}

export interface GetRequestParams {
  path: string;
  type?: DataType;
  data?: Record<string, unknown> | string;
  query?: Record<string, string | number>;
  extraHeaders?: HeaderParams;
  tries?: number;
}

export type PostRequestParams = GetRequestParams & {
  type: DataType;
  data: Record<string, unknown> | string;
};

export type PutRequestParams = PostRequestParams;

export type DeleteRequestParams = GetRequestParams;

export type RequestParams = (GetRequestParams | PostRequestParams) & {
  method: Method;
};

export interface RequestReturn {
  body: unknown;
  headers: Headers;
}